requirements.txt
altair==4.2.2
appdirs==1.4.4
asn1crypto  
attrs==23.1.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.12.2
blinker==1.6.2
Bottleneck 
branca==0.6.0
brotlipy==0.7.0
bs4==0.0.1
cachetools==5.3.1
click==8.1.7
cloudpickle 
contourpy==1.1.1
cryptography 
cycler==0.12.0
entrypoints==0.4
filelock 
folium==0.14.0
fonttools==4.43.0
geographiclib==2.0
geopy==2.3.0
gitdb==4.0.10
GitPython==3.1.37
graphviz==0.20.1
hydralit-components==1.0.10
importlib-metadata==6.8.0
importlib-resources==6.1.0
isodate==0.6.1
Jinja2==3.1.2
joblib 
json2txttree==1.0.4
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
kiwisolver==1.4.5
lxml==4.9.3
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.7.3
matplotlib-venn==0.11.7
mdurl==0.1.2
networkx==3.1
numexpr 
numpy 
oscrypto 
packaging 
pandas 
Pillow==10.0.1
pkgutil_resolve_name==1.3.10
platformdirs 
plotly==5.14.1
pooch 
prettytable==3.9.0
protobuf==3.20.3
pyarrow==10.0.1
pycparser 
pycryptodomex 
pydeck==0.8.1b0
pyecharts==2.0.4
Pygments==2.16.1
PyJSONViewer==1.6.0
PyJWT 
Pympler==1.0.1
pyOpenSSL 
pyparsing==3.1.1
PySocks 
python-dateutil 
python-decouple==3.8
pytz 
PyYAML 
rdflib==7.0.0
referencing==0.30.2
requests 
rich==13.6.0
rpds-py==0.10.3
simplejson==3.19.1
six 
smmap==5.0.1
snowflake-connector-python 
snowflake-snowpark-python 
sortedcontainers 
soupsieve==2.5
SQLAlchemy==2.0.15
streamlit==1.27.2
streamlit-aggrid==0.3.4.post3
streamlit-agraph==0.0.42
streamlit-echarts==0.4.0
streamlit-folium==0.11.1
streamlit-on-Hover-tabs==1.0.1
streamlit-option-menu==0.3.2
tenacity==8.2.3
threadpoolctl 
toml==0.10.2
tomlkit 
toolz==0.12.0
tornado==6.3.3
typing_extensions  
tzdata 
tzlocal==5.0.1
urllib3 
validators==0.22.0
wcwidth==0.2.8
zipp==3.17.0
